mono-jit (#167)
- fix alignment issue with jitcache that could cause a crash.
- Reimplement sparse-jit to only use 1 layer.
    - the old 2 layer approach basically only ever used 1 path due to the bottom-layer size, so doing a double look-up seemed wasteful.
    - changed memory block allocation stategy to allocate based on requested code range instead of the possible address-space.

Reviewed-on: https://git.ryujinx.app/projects/Ryubing/pulls/167

+++ src/ARMeilleure/Translation/TranslatorStubs.cs
@@ -161,7 +161,7 @@ namespace ARMeilleure.Translation
             context.BranchIfTrue(lblFallback, masked);
 
             Operand index = default;
-            Operand page = Const((long)_functionTable.Base);
+            Operand page = Const(_functionTable.Base);
 
             for (int i = 0; i < _functionTable.Levels.Length; i++)
             {
